move files into pages folders
