+++ orienteering/orienteering_frontend/orienteering_frontend.esproj
@@ -12,6 +12,7 @@
     <BuildOutputFolder>$(MSBuildProjectDirectory)\build</BuildOutputFolder>
   </PropertyGroup>
   <ItemGroup>
+    <Folder Include="src\components\" />
     <Folder Include="src\pages\CheckpointDetails\" />
   </ItemGroup>
   <ItemGroup>
