--- orienteering/orienteering_frontend/src/components/DropdownMenu.js
@@ -1,19 +1,0 @@
-//import Dropdown from 'react-bootstrap/Dropdown';
-
-//function DropdownMenu() {
-//    return (
-//        <Dropdown>
-//            <Dropdown.Toggle variant="success" id="dropdown-basic">
-//                Dropdown Button
-//            </Dropdown.Toggle>
-
-//            <Dropdown.Menu>
-//                <Dropdown.Item href="#/action-1">Game1</Dropdown.Item>
-//                <Dropdown.Item href="#/action-2">Game2</Dropdown.Item>
-//                <Dropdown.Item href="#/action-3">Game3</Dropdown.Item>
-//            </Dropdown.Menu>
-//        </Dropdown>
-//    );
-//}
-
-//export default DropdownMenu;
