--- orienteering/orienteering_frontend/src/components/QuizQuestionAlternative.js
@@ -1,4 +1,0 @@
-import { Radio, FormControlLabel } from '@mui/material';
-const QuizQuestionAlternative = ({ data }) => <FormControlLabel value={ data } control={<Radio />} label={data} />
-
-export default QuizQuestionAlternative;
