+++ orienteering/orienteering_frontend/src/pages/QuizPage/QuizQuestionAlternative.js
@@ -1,0 +1,6 @@
+import { Radio, FormControlLabel } from '@mui/material';
+
+
+const QuizQuestionAlternative = ({ data }) => <FormControlLabel value={data} control={<Radio />} label={data} />
+
+export default QuizQuestionAlternative;
