+++ orienteering/orienteering_frontend/src/pages/QuizPage/QuizQuestionItem.js
